modify user info class
